refactor public keys into dev settings
